+++ src/modules/general-affair/entities/bon-makan-khl.vw.ts
@@ -1,0 +1,22 @@
+import { Field, Int, ObjectType } from 'type-graphql';
+import { BaseEntity, Column, Entity, PrimaryColumn } from 'typeorm';
+
+@Entity('ANG_BON_MAKAN_KHL_V')
+@ObjectType()
+export class BonMakanKhlView extends BaseEntity {
+  @Field(() => Int)
+  @PrimaryColumn({ name: 'ID' })
+  id!: number;
+
+  @Field()
+  @Column({ name: 'TANGGAL' })
+  tanggal!: Date;
+
+  @Field()
+  @Column({ name: 'GRADE' })
+  grade!: string;
+
+  @Field()
+  @Column({ name: 'KETERANGAN' })
+  keterangan!: string;
+}